+++ example/main/example_backend.c
@@ -676,7 +676,7 @@ example_statefile(clixon_handle     h,
  * @retval     0   OK
  * @retval    -1   Error
  *
- * System-only config data as defined by _ is not written to datastore.
+ * System-only config data is not written to datastore.
  * Instead, in this ocmmit action, it is written to file _state_file
  * @see main_system_only_commit  callback for reading data
  * @note  Only single system-only config data supported